VmApiProxy
